move no op test to solver file
